rename from /libs to /src
