+++ pages/index.tsx
@@ -5,10 +5,11 @@ import Head from 'next/head'
 import styles from '../styles/Home.module.css'
 import {
   atomRef,
+  useIsNew,
   useReadAtom,
   useSendAtom,
   AtomRoot
-} from '../libs/atomic-state'
+} from '../src/atomic-state'
 
 type $Hello = {
   message: string
@@ -19,7 +20,7 @@ const helloRef = atomRef<$Hello>({
   key: 'Hello',
   defaultState: {
     message: 'Hello world',
-    showTimer: true
+    showTimer: false
   }
 })
 
@@ -66,6 +67,20 @@ const Timer = () => {
   return <div>Time Elapsed: {count}s</div>
 }
 
+const MessageOnly = () => {
+  const { msg } = useReadAtom(
+    helloRef,
+    useIsNew(({ message: msg }) => ({ msg }))
+  )
+
+  return (
+    <div>
+      <h2>Message Only</h2>
+      <div>{msg}</div>
+    </div>
+  )
+}
+
 const AtomAppDemo = () => {
   const text = useReadAtom(helloRef, (s) => s.message)
   const showTimer = useReadAtom(
@@ -76,6 +91,7 @@ const AtomAppDemo = () => {
 
   return (
     <div>
+      <h2>Atom App Main Demo</h2>
       <div>
         <input
           type="text"
@@ -107,7 +123,7 @@ const Home: NextPage = () => {
     () =>
       dynamic(
         async () =>
-          (await import('../libs/atomic-state'))
+          (await import('../src/atomic-state'))
             .AtomDevTools,
         { ssr: false }
       ),
@@ -129,6 +145,7 @@ const Home: NextPage = () => {
         <h1>React Atomic</h1>
 
         <AtomRoot>
+          <MessageOnly />
           <AtomAppDemo />
           <AtomDevTools />
         </AtomRoot>
