+++ src/atomic-state/constants.ts
